foundry-agent-v2-python-deployment/requirements.txt
azure.ai.agents
azure.identity
dotenv
jinja2
azure-ai-projects>=2.0.0